Fixed typing

+++ occurrence-download-launcher/src/main/java/org/gbif/occurrence/downloads/launcher/pojo/SparkStaticConfiguration.java
@@ -67,5 +67,5 @@ public class SparkStaticConfiguration {
 
   // Making it possible to configure the value via configurations
   private int vectorMemory;
-  private double vectorCpu;
+  private int vectorCpu;
 }
